+++ src/operations.h
@@ -38,8 +38,8 @@ namespace sharp {
   /*
    * Convolution with a kernel.
    */
-  VImage Convolve(VImage image, int width, int height, double scale, double offset,
-                  const std::unique_ptr<double[]> &kernel_v);
+  VImage Convolve(VImage image, int const width, int const height,
+    double const scale, double const offset, std::unique_ptr<double[]> const &kernel_v);
 
   /*
    * Sharpen flat and jagged areas. Use sigma of -1.0 for fast sharpen.